Fix fullscreen in xdg-shell
